Refactor overview into verfahren route
- Show Verfahren overview instantly after login
- No need to show an additional overview/dashboard route at this moment
